Move StateManager
